Add subset_bamfile, improve speed and memory usage.
We use a buffered IO reader to speed up opening the gzipped
FASTQ file. This speeds up the reading by an order of magnitude.

We also lowered memory usage of tagcount by an order of magnitude by
explicitely deleting objects we are not using anymore and writing the evidence
table out to a temporary file before loading it with pandas.