Add `fasttagcount` and `sparse` subcommands.
`fasttagcount` assumes the input BAM/SAM file is coordinate sorted,
and only keeps the counts for a single transcript at a time in
memory. This results in a much smaller memory footprint for
deep datasets, went from > 100 GB to 1 GB for an example dataset.
It is also 30-40% faster. This should maybe just replace
tagcount if everyone is on board.

`sparse` takes a regular matrix and outputs a sparse matrix.
This is to offload the option creep going into some of the other
functions.

This commit also adds tests for the `--genemap` option, and
ensures that the index column is labelled 'gene' when the
gene-map is used.